--- app/src/main/java/com/epfl/beatlink/model/spotify/MusicServiceAuthRepository.kt
@@ -1,3 +1,0 @@
-package com.epfl.beatlink.model.spotify
-
-interface MusicServiceAuthRepository
