+++ app/src/main/java/com/epfl/beatlink/model/spotify/auth/MusicServiceAuthRepository.kt
@@ -1,0 +1,3 @@
+package com.epfl.beatlink.model.spotify.auth
+
+interface MusicServiceAuthRepository
