+++ app/src/test/java/com/epfl/beatlink/model/spotify/api/SpotifyApiRepositoryTest.kt
@@ -1,0 +1,3 @@
+package com.epfl.beatlink.model.spotify.api
+
+class SpotifyApiRepositoryTest {}
